Update co2snatcher.ino
This code was originally coded with STM32CubeIDE, but due to technical difficulties, the code is malfunctioning. Instead we  installed STM32 add-on program on Arduino so that the board can be programmed with Arduino for better performance. By using the MQ2 smoke detector connected to the STM32 microcontroller, this code detects the CO2 level (ppml) in the air and displays it in real time on the LCD. We include two libraries at the beginning of the code, the MQ2 library and the LCD display library, and then we set the required variables and LCD display size. Lastly, we set the smoke detector switch and display on in the void set function so that the whole project can be powered on. In the loop, we converted CO2 levels in the air to ppml using the method provided in the MQ2 smoke detector library. The delay time was set so that the data could be displayed on the LCD for 10 seconds, which is more convenient for logging.